Refactor backend to align with architectural principles
