+++ functions/ai/dossier_generator.py
